+++ code/include/FixPolicy.hpp
@@ -4,8 +4,8 @@
  *        specific policy-based fixing strategies in the ACS framework.
  *
  * @author Francesco Biscaccia Carrara
- * @version v1.2.5
- * @since 06/21/2025
+ * @version v1.2.6
+ * @since 06/23/2025
  */
 
 #ifndef FIX_POL_H