+++ code/include/MIP.hpp
@@ -8,8 +8,8 @@
  *
  * @note Requires CPLEX library and Utils.hpp
  * @author Francesco Biscaccia Carrara
- * @version v1.2.5
- * @since 06/21/2025
+ * @version v1.2.6
+ * @since 06/23/2025
  */
 
 #ifndef MIP_SOL_H
@@ -29,7 +29,6 @@ using namespace Utils;
 #define NUM_SOL_STOP 1
 //#define ACS_CB_CONTEXTMASK CPX_CALLBACKCONTEXT_GLOBAL_PROGRESS | CPX_CALLBACKCONTEXT_CANDIDATE
 
-// FIXME: can they be removed?
 #define MIP_DUAL_PRIM_GAP_TOL 1e-4 // Default value for CPX_PARAM_EPAGAP (see IBM ILOG CPLEX doc)
 #define MIP_GAP_TOL 1e-6		   // Default value for CPX_PARAM_EPGAP (see IBM ILOG CPLEX doc)
 #define MIP_INT_TOL 1e-3